refactor(fzf.vim): majorly improve fzf/rg search
Setup commands keybindings for the following:
- file search of current directory
- file search from the git root
- file content search from the git root
- file content search of the current directory retriggering ripgrep
  instead of filtering with fzf (for large folders/repos)
- commit search for the whole git repository
- commit search for the current file/buffer

As there is now a differentiation between search the current directory
and from the git root, this allows us to reenable "autochdir" to switch
to the directory of the current file.
So now it is possible to search through the whole project but also open
files and search files relative to the current location.
Best of both worlds!